requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
Pillow==9.4.0
cryptography==39.0.1